backend/requirements.txt
gunicorn
hupper
numpy
psycopg2-binary
pytest
botocore
aws-secretsmanager-caching